Grid Frozen Conditions now active in FK grid combo relcol selector
API change

For the past 11 years, saved grid views when used as grid-combo for
FK/single-rel selector had filters cleared because it was deemed to
cause more user confusion than benefit. But not doing this is now a
greater desire in certain cases. To break back-compat less, this is now
turned off only for Frozen multifilters, making the feature available
but in a sub-set of scenarios rather than making an across-the-board
change

This situation arises mainly when NavCore is enabled and a "Source
Default View" is chosen for a given Result Source.

Note that normal Filters And Quick Search in saved views are still
cleared in this scenario