Use either _resolve_relationship_condition or resolve_relationship_condition
Newer versions of DBIC added a new, more powerful method called
'resolve_relationship_condition' which has extended and given fuller
capabilities over the previous '_resolve_condition' method, which could
not handle all cases, most notably, relationships which have 'where'
clause attrs defined, which is something we (RapidApp) need/want to
support for multi-relationship columns.

This issue was addressed a good while ago before the first DBIC release
of the new method into the wild, and code was added to support both
cases, checking for the existence of the method. It seems that since
that time, DBIC did not make a release officially adding
'resolve_relationship_condition' after all, but did make a release with
'_resolve_relationship_condition', which by convention denotes that it
should be considered "private". I didn't actually realize that this
happened this way, and so the old, legacy code path using
'_resolve_condition' has continued being used in all cases of multi rels
with a 'where' not working properly, and also outright breaking
(throwing exception) on relationships with a 'where' which also use a
coderef condition, and it was this latter case that brought this
situation to light.

The code now checks for both resolve_relationship_condition and
_resolve_relationship_condition, preferentially using the former if it
exists, so if/when this is finally exposed as a public-named method, it
will be using the more correct API. If you're a RapidApp maintainer
reading this in the future, and DBIC has since released the
final/public-named method, you should unfactor all this fall-back code
and use only '->resolve_relationship_condition()' and require the most
recent version of DBIC.